backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.6.1
httpx==0.24.1
email-validator==2.1.0
azure-storage-blob==12.19.0
python-multipart==0.0.6
